requirements.txt
streamlit 
google-cloud-storage 
pandas 
seleniumbase 
beautifulsoup4
pandas